feat(exif): split exif document classes
