+++ src/Model/Exif/ExifDocument.php
@@ -14,34 +14,18 @@ namespace MagicSunday\ImageMeta\Model\Exif;
 use DateTimeImmutable;
 use DateTimeZone;
 
-final class IfdEntry
-{
-    public function __construct(
-        public readonly int $tag,
-        public readonly int $type,
-        public readonly int $count,
-        public readonly mixed $value,
-    ) {
-    }
-}
-
-final class Ifd
-{
-    /** @param array<int, IfdEntry> $entries */
-    public function __construct(
-        public readonly array $entries,
-        public readonly ?int $nextIfdOffset = null,
-    ) {
-    }
-
-    public function get(int $tag): ?IfdEntry
-    {
-        return $this->entries[$tag] ?? null;
-    }
-}
-
+/**
+ * Represents a parsed EXIF payload and exposes convenience accessors.
+ */
 final class ExifDocument
 {
+    /**
+     * @param Ifd      $ifd0       Root IFD of the TIFF structure.
+     * @param Ifd|null $exifIfd    Sub IFD containing EXIF-specific tags.
+     * @param Ifd|null $gpsIfd     Sub IFD containing GPS-related tags.
+     * @param Ifd|null $interopIfd Sub IFD containing interoperability tags.
+     * @param Ifd|null $ifd1       Optional next IFD, typically thumbnails.
+     */
     public function __construct(
         public readonly Ifd $ifd0,
         public readonly ?Ifd $exifIfd,
@@ -51,33 +35,62 @@ final class ExifDocument
     ) {
     }
 
-    // ---- Convenience (Strings/Numbers)
+    /**
+     * Returns the camera manufacturer string if present.
+     *
+     * @return string|null
+     */
     public function cameraMake(): ?string
     {
         return $this->str($this->ifd0, 0x010F);
     }
 
+    /**
+     * Returns the camera model string if present.
+     *
+     * @return string|null
+     */
     public function cameraModel(): ?string
     {
         return $this->str($this->ifd0, 0x0110);
     }
 
+    /**
+     * Returns the lens model string if present.
+     *
+     * @return string|null
+     */
     public function lensModel(): ?string
     {
         return $this->str($this->exifIfd, 0xA434);
     }
 
+    /**
+     * Returns the EXIF orientation value if present.
+     *
+     * @return int|null
+     */
     public function orientation(): ?int
     {
         return $this->int($this->ifd0, 0x0112);
     }
 
+    /**
+     * Returns the ISO sensitivity value if present.
+     *
+     * @return int|null
+     */
     public function iso(): ?int
     {
         // EXIF ISO tag (PhotographicSensitivity) 0x8827
         return $this->int($this->exifIfd, 0x8827);
     }
 
+    /**
+     * Returns the exposure time in seconds if available.
+     *
+     * @return float|null
+     */
     public function exposureTime(): ?float
     {
         // Tag 0x829A RATIONAL
@@ -86,6 +99,11 @@ final class ExifDocument
         return ValueConverters::rationalToFloat($v);
     }
 
+    /**
+     * Returns the aperture (f-number) if available.
+     *
+     * @return float|null
+     */
     public function fNumber(): ?float
     {
         // Tag 0x829D RATIONAL
@@ -94,6 +112,11 @@ final class ExifDocument
         return ValueConverters::rationalToFloat($v);
     }
 
+    /**
+     * Returns the focal length in millimetres if available.
+     *
+     * @return float|null
+     */
     public function focalLengthMm(): ?float
     {
         // Tag 0x920A RATIONAL
@@ -102,17 +125,31 @@ final class ExifDocument
         return ValueConverters::rationalToFloat($v);
     }
 
+    /**
+     * Returns the raw DateTimeOriginal tag value.
+     *
+     * @return string|null
+     */
     public function dateTimeOriginalRaw(): ?string
     {
         return $this->str($this->exifIfd, 0x9003);
     }
 
+    /**
+     * Returns the raw offset time for DateTimeOriginal.
+     *
+     * @return string|null
+     */
     public function offsetTimeOriginalRaw(): ?string
     {
         return $this->str($this->exifIfd, 0x9011);
     }
 
-    /** @return array{lat:?float, lon:?float, alt:?float} */
+    /**
+     * Returns the parsed GPS coordinates if present.
+     *
+     * @return array{lat:?float, lon:?float, alt:?float}
+     */
     public function gps(): array
     {
         if (!$this->gpsIfd) {
@@ -122,7 +159,11 @@ final class ExifDocument
         return ValueConverters::gpsFromIfd($this->gpsIfd);
     }
 
-    /** Best-effort CaptureTime (DateTimeImmutable). Falls kein Offset vorhanden, Timezone = UTC. */
+    /**
+     * Returns a best-effort capture timestamp. Defaults to UTC when no offset tag is provided.
+     *
+     * @return DateTimeImmutable|null
+     */
     public function captureDateTime(): ?DateTimeImmutable
     {
         $raw = $this->dateTimeOriginalRaw();
@@ -139,6 +180,11 @@ final class ExifDocument
         return $dt ?: null;
     }
 
+    /**
+     * Returns a string value from the given IFD if present.
+     *
+     * @return string|null
+     */
     private function str(?Ifd $ifd, int $tag): ?string
     {
         $e = $ifd?->get($tag);
@@ -146,64 +192,15 @@ final class ExifDocument
         return is_string($e?->value) ? rtrim($e->value, "\0") : null;
     }
 
+    /**
+     * Returns an integer value from the given IFD if present.
+     *
+     * @return int|null
+     */
     private function int(?Ifd $ifd, int $tag): ?int
     {
         $v = $ifd?->get($tag)?->value ?? null;
 
         return is_int($v) ? $v : (is_float($v) ? (int) $v : null);
-    }
-}
-
-final class ValueConverters
-{
-    public static function rationalToFloat(mixed $v): ?float
-    {
-        if (is_array($v) && count($v) === 2 && (int) $v[1] !== 0) {
-            return (float) $v[0] / (float) $v[1];
-        }
-        if (is_int($v) || is_float($v)) {
-            return (float) $v;
-        }
-
-        return null;
-    }
-
-    /** @return array{lat:?float,lon:?float,alt:?float} */
-    public static function gpsFromIfd(Ifd $gps): array
-    {
-        $latRef = $gps->get(0x0001)?->value ?? null;
-        $latVal = $gps->get(0x0002)?->value ?? null;
-        $lonRef = $gps->get(0x0003)?->value ?? null;
-        $lonVal = $gps->get(0x0004)?->value ?? null;
-
-        $lat = self::dmsToFloat($latRef, $latVal);
-        $lon = self::dmsToFloat($lonRef, $lonVal);
-
-        $alt = null;
-        if (($e = $gps->get(0x0006)) && is_array($e->value) && count($e->value) === 2 && (int) $e->value[1] !== 0) {
-            $alt = $e->value[0] / $e->value[1];
-            if (($ref = $gps->get(0x0005)) && (int) ($ref->value ?? 0) === 1) {
-                $alt = -$alt;
-            }
-        }
-
-        return ['lat' => $lat, 'lon' => $lon, 'alt' => $alt];
-    }
-
-    private static function dmsToFloat(?string $ref, mixed $val): ?float
-    {
-        if (!is_string($ref) || !is_array($val) || count($val) < 3) {
-            return null;
-        }
-        $deg = self::rationalToFloat($val[0] ?? null);
-        $min = self::rationalToFloat($val[1] ?? null);
-        $sec = self::rationalToFloat($val[2] ?? null);
-        if ($deg === null || $min === null || $sec === null) {
-            return null;
-        }
-
-        $sign = ($ref === 'S' || $ref === 'W') ? -1.0 : 1.0;
-
-        return $sign * ($deg + $min / 60.0 + $sec / 3600.0);
     }
 }
